gzgets should take an unsigned integer, not a signed one
Every usage of this function intends on passing an unsigned integer anyway, as they do in gzread or gzwrite.

I found bugs where to prevent overflow before calling this function, programmers checked for UINT_MAX, and not INT_MAX.

+++ test/example.c
@@ -176,7 +176,7 @@ void test_gzio(fname, uncompr, uncomprLen)
         exit(1);
     }
 
-    gzgets(file, (char*)uncompr, (int)uncomprLen);
+    gzgets(file, (char*)uncompr, uncomprLen);
     if (strlen((char*)uncompr) != 7) { /* " hello!" */
         fprintf(stderr, "gzgets err after gzseek: %s\n", gzerror(file, &err));
         exit(1);
